Fix: Errores de componentes y diseño del buscador

+++ server/src/models/careers.js
@@ -20,7 +20,7 @@ export const Career = sequelize.define('Career', {
         allowNull: false,
     },
     id_institute: {
-        type: DataTypes.INTEGER(1),
+        type: DataTypes.INTEGER(4),
         allowNull: true,
     },
     name: {
